test: extract LegendList Jest mock
